disasm: Add explicit type conversion to placate compiler
Otherwise getting

 | disasm.c:200:26: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
 |         return GET_REGISTER(nasm_rd_bndreg, regval);

+++ disasm.c
@@ -159,7 +159,7 @@ static enum reg_enum whichreg(opflags_t regflags, int regval, int rex)
         return 0;
 
 #define GET_REGISTER(__array, __index)                      \
-    ((__index) < ARRAY_SIZE(__array) ? __array[(__index)] : 0)
+    ((size_t)(__index) < (size_t)ARRAY_SIZE(__array) ? __array[(__index)] : 0)
 
     if (!(REG8 & ~regflags)) {
         if (rex & (REX_P|REX_NH))
